feat: mbti 타입 조회, 문항 조회, 결과 조회, 결과 테스트 api 구현

+++ src/main/java/org/umc/valuedi/domain/mbti/dto/FinanceMbtiTestRequestDto.java
@@ -13,6 +13,6 @@ public record FinanceMbtiTestRequestDto(
 ) {
     public record Answer(
             @NotNull Long questionId,
-            @NotNull @Min(1) @Max(5) Byte choiceValue
+            @NotNull @Min(1) @Max(5) Integer choiceValue
     ) {}
 }
